Refactor(dam_archive): Split commands and systems into submodules
This refactoring splits the `commands.py` and `systems.py` modules in the `dam_archive` package into smaller, more focused submodules based on their functionality. The new structure is as follows:

- `dam_archive.commands`:
  - `ingestion.py`
  - `password.py`
  - `split_archives.py`
- `dam_archive.systems`:
  - `discovery.py`
  - `ingestion.py`
  - `password.py`
  - `split_archives.py`

Additionally, the deprecated `DiscoverAndBindCommand` and its corresponding handler, `discover_and_bind_handler`, have been removed from the codebase.

All imports have been updated to reflect the new module structure, and the tests have been updated and verified to pass after the changes.

Also includes:
- Update project dependencies to the latest versions.
- Update `run_task_in_pkgs_if_exist.py` to correctly handle passing arguments to underlying tasks like pytest.
- Update `README.md` and `AGENTS.md` with instructions on how to run specific tests using pytest, now supported by the updated script.
