+++ packages/dam_archive/src/dam_archive/commands/__init__.py
@@ -1,0 +1,33 @@
+# Main entry point for commands in the dam_archive package.
+
+from .ingestion import (
+    CheckArchiveCommand,
+    ClearArchiveComponentsCommand,
+    IngestArchiveCommand,
+    TagArchivePartCommand,
+)
+from .password import (
+    CheckArchivePasswordCommand,
+    RemoveArchivePasswordCommand,
+    SetArchivePasswordCommand,
+)
+from .split_archives import (
+    BindSplitArchiveCommand,
+    CheckSplitArchiveBindingCommand,
+    CreateMasterArchiveCommand,
+    UnbindSplitArchiveCommand,
+)
+
+__all__ = [
+    "CheckArchiveCommand",
+    "ClearArchiveComponentsCommand",
+    "IngestArchiveCommand",
+    "TagArchivePartCommand",
+    "CheckArchivePasswordCommand",
+    "RemoveArchivePasswordCommand",
+    "SetArchivePasswordCommand",
+    "BindSplitArchiveCommand",
+    "CheckSplitArchiveBindingCommand",
+    "CreateMasterArchiveCommand",
+    "UnbindSplitArchiveCommand",
+]
