+++ packages/dam_archive/src/dam_archive/commands/ingestion.py
@@ -1,0 +1,51 @@
+from dataclasses import dataclass
+from typing import Dict, List, Optional
+
+from dam.commands.analysis_commands import AnalysisCommand
+from dam.commands.core import BaseCommand, EntityCommand
+from dam.system_events.base import BaseSystemEvent
+
+
+@dataclass
+class IngestArchiveCommand(AnalysisCommand[None, BaseSystemEvent]):
+    """
+    A command to ingest members from an archive asset into the ECS world.
+    This command returns a stream of events.
+    """
+
+    passwords: Optional[List[str]] = None
+
+    @classmethod
+    def get_supported_types(cls) -> Dict[str, List[str]]:
+        """
+        Returns a dictionary of supported MIME types and file extensions for archives.
+        """
+        return {
+            "mimetypes": [],
+            "extensions": [".zip", ".rar", ".7z"],
+        }
+
+
+@dataclass
+class TagArchivePartCommand(BaseCommand[None, BaseSystemEvent]):
+    """
+    A command to tag a file as a potential split archive part.
+    """
+
+    entity_id: int
+
+
+@dataclass
+class CheckArchiveCommand(EntityCommand[bool, BaseSystemEvent]):
+    """A command to check if an entity has been processed as an archive."""
+
+    pass
+
+
+@dataclass
+class ClearArchiveComponentsCommand(EntityCommand[None, BaseSystemEvent]):
+    """
+    A command to clear archive-related components from an entity.
+    """
+
+    pass
