+++ packages/dam_archive/src/dam_archive/systems/ingestion.py
@@ -1,9 +1,6 @@
 import io
 import logging
-import os
 from contextlib import asynccontextmanager
-from datetime import datetime, timezone
-from pathlib import Path
 from typing import (
     Annotated,
     AsyncContextManager,
@@ -13,7 +10,6 @@ from typing import (
     Dict,
     List,
     Optional,
-    Tuple,
     Union,
     cast,
 )
@@ -24,13 +20,11 @@ from dam.commands.asset_commands import (
     GetAssetStreamCommand,
     GetOrCreateEntityFromStreamCommand,
 )
-from dam.commands.discovery_commands import DiscoverPathSiblingsCommand, PathSibling
 from dam.core.systems import system
 from dam.core.transaction import WorldTransaction
 from dam.core.types import CallableStreamProvider, StreamProvider
 from dam.core.world import World
 from dam.functions.mime_type_functions import get_content_mime_type
-from dam.models.metadata.content_mime_type_component import ContentMimeTypeComponent
 from dam.system_events.entity_events import NewEntityCreatedEvent
 from dam.system_events.progress import (
     ProgressCompleted,
@@ -40,27 +34,17 @@ from dam.system_events.progress import (
     SystemProgressEvent,
 )
 from dam.utils.stream_utils import ChainedStream
-from dam_fs.commands import FindEntityByFilePropertiesCommand
-from dam_fs.models import FilenameComponent
 from sqlalchemy import select
 
-from . import split_detector
-from .commands import (
-    BindSplitArchiveCommand,
+from ..commands.ingestion import (
     CheckArchiveCommand,
-    CheckArchivePasswordCommand,
-    CheckSplitArchiveBindingCommand,
     ClearArchiveComponentsCommand,
-    CreateMasterArchiveCommand,
-    DiscoverAndBindCommand,
     IngestArchiveCommand,
-    RemoveArchivePasswordCommand,
-    SetArchivePasswordCommand,
-    UnbindSplitArchiveCommand,
 )
-from .exceptions import InvalidPasswordError, PasswordRequiredError
-from .main import open_archive
-from .models import (
+from ..commands.password import SetArchivePasswordCommand
+from ..exceptions import InvalidPasswordError, PasswordRequiredError
+from ..main import open_archive
+from ..models import (
     ArchiveInfoComponent,
     ArchiveMemberComponent,
     ArchivePasswordComponent,
@@ -71,47 +55,6 @@ from .models import (
 logger = logging.getLogger(__name__)
 
 
-@system(on_command=DiscoverPathSiblingsCommand)
-async def discover_archive_path_siblings_handler(
-    cmd: DiscoverPathSiblingsCommand,
-    transaction: WorldTransaction,
-) -> Optional[List[PathSibling]]:
-    """
-    Handles discovering path-based sibling entities for an entity that is a member of an archive.
-    """
-    logger.debug(f"discover_archive_path_siblings_handler running for entity {cmd.entity_id}")
-
-    # 1. Get the ArchiveMemberComponent for the starting entity
-    member_comp = await transaction.get_component(cmd.entity_id, ArchiveMemberComponent)
-    if not member_comp or not member_comp.path_in_archive:
-        logger.debug(f"Entity {cmd.entity_id} is not an archive member. Skipping archive discovery.")
-        return None
-
-    # 2. Determine the parent archive and the directory within it
-    parent_archive_id = member_comp.archive_entity_id
-    internal_dir = os.path.dirname(member_comp.path_in_archive)
-
-    # 3. Find all members of the same parent archive first.
-    stmt = select(ArchiveMemberComponent).where(ArchiveMemberComponent.archive_entity_id == parent_archive_id)
-    result = await transaction.session.execute(stmt)
-    all_members = result.scalars().all()
-
-    # 4. Filter them by directory and create PathSibling objects
-    siblings = [
-        PathSibling(entity_id=m.entity_id, path=m.path_in_archive)
-        for m in all_members
-        if m.path_in_archive and os.path.dirname(m.path_in_archive) == internal_dir
-    ]
-
-    if siblings:
-        logger.info(
-            f"Found {len(siblings)} archive siblings for entity {cmd.entity_id} in archive {parent_archive_id}:{internal_dir}."
-        )
-        return siblings
-
-    return None
-
-
 @system(on_command=CheckArchiveCommand)
 async def check_archive_handler(
     cmd: CheckArchiveCommand,
@@ -120,326 +63,6 @@ async def check_archive_handler(
     """Checks if the ArchiveInfoComponent exists for the entity."""
     component = await transaction.get_component(cmd.entity_id, ArchiveInfoComponent)
     return component is not None
-
-
-@system(on_command=DiscoverAndBindCommand)
-async def discover_and_bind_handler(
-    cmd: DiscoverAndBindCommand,
-    transaction: WorldTransaction,
-    world: Annotated[World, "Resource"],
-):
-    """
-    Scans given paths, detects complete split archives, and creates a master entity for each.
-    """
-    logger.info(f"Starting discovery and binding for paths: {cmd.paths}")
-
-    # 1. Scan file system and detect split archive parts
-    all_files: List[Tuple[str, float]] = []
-    for p_str in cmd.paths:
-        p = Path(p_str)
-        if not p.exists():
-            logger.warning(f"Path does not exist, skipping: {p}")
-            continue
-        if p.is_dir():
-            for root, _, files in os.walk(p):
-                for name in files:
-                    full_path = os.path.join(root, name)
-                    try:
-                        mtime = os.path.getmtime(full_path)
-                        all_files.append((full_path, mtime))
-                    except FileNotFoundError:
-                        continue
-        else:
-            try:
-                mtime = os.path.getmtime(p)
-                all_files.append((str(p), mtime))
-            except FileNotFoundError:
-                pass
-
-    # Group detected parts by base_name
-    parts_by_basename: Dict[str, List[Tuple[str, float, int]]] = {}
-    for path, mtime in all_files:
-        split_info = split_detector.detect(os.path.basename(path))
-        if split_info:
-            parts_by_basename.setdefault(split_info.base_name, []).append((path, mtime, split_info.part_num))
-
-    # 2. Process each group to find complete archives
-    for base_name, parts_with_meta in parts_by_basename.items():
-        parts_by_num = {part_num: (path, mtime) for path, mtime, part_num in parts_with_meta}
-        if not parts_by_num:
-            continue
-
-        max_part_num = max(parts_by_num.keys())
-        is_complete = all(i in parts_by_num for i in range(1, max_part_num + 1))
-
-        if is_complete:
-            logger.info(f"Found complete split archive '{base_name}' with {max_part_num} parts.")
-
-            part_entity_ids: List[int] = []
-            is_valid_group = True
-
-            # 3. Find entity IDs for each part
-            sorted_parts_meta = [parts_by_num[i] for i in range(1, max_part_num + 1)]
-
-            for path, mtime in sorted_parts_meta:
-                file_uri = Path(path).as_uri()
-                modified_at = datetime.fromtimestamp(mtime, tz=timezone.utc)
-
-                # Dispatch command to find entity
-                find_cmd = FindEntityByFilePropertiesCommand(file_path=file_uri, last_modified_at=modified_at)
-                try:
-                    entity_id = await world.dispatch_command(find_cmd).get_one_value()
-                except ValueError:
-                    entity_id = None
-
-                if entity_id:
-                    part_entity_ids.append(entity_id)
-                else:
-                    logger.warning(
-                        f"Could not find a matching entity for part '{path}' with mtime {mtime}. "
-                        "Skipping assembly for this group."
-                    )
-                    is_valid_group = False
-                    break
-
-            # 4. Dispatch CreateMasterArchiveCommand
-            if is_valid_group:
-                logger.info(f"Assembling master archive for '{base_name}' with parts: {part_entity_ids}")
-                master_name = f"{base_name} (Split Archive)"
-                create_cmd = CreateMasterArchiveCommand(name=master_name, part_entity_ids=part_entity_ids)
-                async for _ in world.dispatch_command(create_cmd):
-                    pass
-        else:
-            logger.info(f"Split archive '{base_name}' is incomplete. Skipping assembly.")
-
-
-@system(on_command=CreateMasterArchiveCommand)
-async def create_master_archive_handler(
-    cmd: CreateMasterArchiveCommand,
-    transaction: WorldTransaction,
-):
-    """
-    Handles the manual creation of a master entity for a split archive.
-    """
-    logger.info(f"Manually creating master archive '{cmd.name}' for {len(cmd.part_entity_ids)} parts.")
-
-    # 1. Create master entity and its components
-    master_entity = await transaction.create_entity()
-    await transaction.add_component_to_entity(
-        master_entity.id,
-        FilenameComponent(filename=cmd.name, first_seen_at=datetime.now(timezone.utc)),
-    )
-    manifest = SplitArchiveManifestComponent()
-    await transaction.add_component_to_entity(master_entity.id, manifest)
-
-    # 2. Copy mime type from first part
-    if cmd.part_entity_ids:
-        first_part_id = cmd.part_entity_ids[0]
-        content_mime_comp = await transaction.get_component(first_part_id, ContentMimeTypeComponent)
-        if content_mime_comp:
-            await transaction.add_component_to_entity(
-                master_entity.id,
-                ContentMimeTypeComponent(mime_type_concept_id=content_mime_comp.mime_type_concept_id),
-            )
-
-    # 3. Update all part components to link to the new master
-    for part_id in cmd.part_entity_ids:
-        fnc = await transaction.get_component(part_id, FilenameComponent)
-        if not fnc or not fnc.filename:
-            logger.warning(f"Skipping part entity {part_id} as it has no filename component.")
-            continue
-
-        split_info = split_detector.detect(fnc.filename)
-        if not split_info:
-            logger.warning(f"Skipping part entity {part_id} as it does not look like a split archive part.")
-            continue
-
-        part_info_comp = await transaction.get_component(part_id, SplitArchivePartInfoComponent)
-        if part_info_comp:
-            part_info_comp.master_entity_id = master_entity.id
-        else:
-            new_part_info = SplitArchivePartInfoComponent(
-                part_num=split_info.part_num,
-                master_entity_id=master_entity.id,
-            )
-            await transaction.add_component_to_entity(part_id, new_part_info)
-
-    logger.info(f"Successfully created master entity {master_entity.id} for archive '{cmd.name}'.")
-
-
-@system(on_command=BindSplitArchiveCommand)
-async def bind_split_archive_handler(
-    cmd: BindSplitArchiveCommand,
-    world: Annotated[World, "Resource"],
-    transaction: WorldTransaction,
-):
-    """
-    Handles discovering and binding a split archive from a starting entity.
-    This is the "consumer" of the generic sibling discovery.
-    """
-    logger.info(f"Attempting to bind split archive starting from entity {cmd.entity_id}")
-
-    # 1. Discover sibling entities
-    discover_cmd = DiscoverPathSiblingsCommand(entity_id=cmd.entity_id)
-    try:
-        siblings = await world.dispatch_command(discover_cmd).get_first_non_none_value()
-    except ValueError:
-        siblings = None
-
-    if not siblings:
-        logger.info(f"No siblings found for entity {cmd.entity_id}. Cannot bind split archive.")
-        return
-
-    # 2. Use the split_detector to find and validate a complete group from the siblings
-    parts_by_basename: Dict[str, Dict[int, int]] = {}  # {base_name: {part_num: entity_id}}
-    for sibling in siblings:
-        split_info = split_detector.detect(os.path.basename(sibling.path))
-        if split_info:
-            parts_by_basename.setdefault(split_info.base_name, {})[split_info.part_num] = sibling.entity_id
-
-    # 3. Process the groups to find a complete one
-    for base_name, parts_by_num in parts_by_basename.items():
-        if not parts_by_num:
-            continue
-
-        max_part_num = max(parts_by_num.keys())
-        is_complete = all(i in range(1, max_part_num + 1) for i in parts_by_num)
-
-        if is_complete:
-            logger.info(f"Found complete split archive '{base_name}' with {max_part_num} parts.")
-
-            # Check if a master archive already exists for this group
-            master_name = f"{base_name} (Split Archive)"
-            stmt = select(FilenameComponent).where(FilenameComponent.filename == master_name)
-            result = await transaction.session.execute(stmt)
-            if result.scalar_one_or_none():
-                logger.warning(f"A master archive named '{master_name}' already exists. Skipping creation.")
-                continue
-
-            # 4. Dispatch command to create the master entity
-            part_entity_ids = [parts_by_num[i] for i in range(1, max_part_num + 1)]
-            create_cmd = CreateMasterArchiveCommand(
-                name=master_name,
-                part_entity_ids=part_entity_ids,
-            )
-            async for _ in world.dispatch_command(create_cmd):
-                pass
-
-            # We only bind the first complete group we find
-            return
-
-
-@system(on_command=CheckSplitArchiveBindingCommand)
-async def check_split_archive_binding_handler(
-    cmd: CheckSplitArchiveBindingCommand,
-    transaction: WorldTransaction,
-) -> bool:
-    """
-    Checks if an entity is part of a fully bound split archive.
-    """
-    # Case 1: The entity is the master archive itself.
-    manifest = await transaction.get_component(cmd.entity_id, SplitArchiveManifestComponent)
-    if manifest:
-        return True
-
-    # Case 2: The entity is a part of an archive.
-    part_info = await transaction.get_component(cmd.entity_id, SplitArchivePartInfoComponent)
-    if part_info and part_info.master_entity_id:
-        # Check if the master it points to is still valid.
-        master_manifest = await transaction.get_component(part_info.master_entity_id, SplitArchiveManifestComponent)
-        if master_manifest:
-            return True
-
-    return False
-
-
-@system(on_command=UnbindSplitArchiveCommand)
-async def unbind_split_archive_handler(
-    cmd: UnbindSplitArchiveCommand,
-    transaction: WorldTransaction,
-):
-    """
-    Handles unbinding a split archive, starting from either a master or a part.
-    """
-    master_entity_id: Optional[int] = None
-
-    # Determine the master entity ID
-    # Case 1: The command is run on the master entity itself.
-    manifest = await transaction.get_component(cmd.entity_id, SplitArchiveManifestComponent)
-    if manifest:
-        master_entity_id = cmd.entity_id
-    # Case 2: The command is run on a part entity.
-    else:
-        part_info = await transaction.get_component(cmd.entity_id, SplitArchivePartInfoComponent)
-        if part_info and part_info.master_entity_id:
-            master_entity_id = part_info.master_entity_id
-
-    if not master_entity_id:
-        logger.warning(f"Entity {cmd.entity_id} is not a split archive master or part. Nothing to unbind.")
-        return
-
-    logger.info(f"Unbinding split archive master entity {master_entity_id}.")
-
-    # Get the manifest component of the actual master
-    master_manifest = await transaction.get_component(master_entity_id, SplitArchiveManifestComponent)
-    if not master_manifest:
-        # This case could happen if the master was deleted but the part component remained.
-        logger.error(
-            f"Could not find a split archive manifest for master entity {master_entity_id}, though part {cmd.entity_id} referenced it."
-        )
-        return
-
-    # Delete part info from all parts associated with the master
-    stmt = select(SplitArchivePartInfoComponent).where(
-        SplitArchivePartInfoComponent.master_entity_id == master_entity_id
-    )
-    result = await transaction.session.execute(stmt)
-    parts_to_unbind = result.scalars().all()
-    for part_info in parts_to_unbind:
-        await transaction.remove_component(part_info)
-
-    # Delete the manifest from the master
-    await transaction.remove_component(master_manifest)
-
-    logger.info(f"Successfully unbound archive for master entity {master_entity_id}.")
-
-
-@system(on_command=CheckArchivePasswordCommand)
-async def check_archive_password_handler(
-    cmd: CheckArchivePasswordCommand,
-    transaction: WorldTransaction,
-) -> bool:
-    """Checks if the ArchivePasswordComponent exists for the entity."""
-    component = await transaction.get_component(cmd.entity_id, ArchivePasswordComponent)
-    return component is not None
-
-
-@system(on_command=RemoveArchivePasswordCommand)
-async def remove_archive_password_handler(
-    cmd: RemoveArchivePasswordCommand,
-    transaction: WorldTransaction,
-):
-    """Removes the ArchivePasswordComponent from the entity."""
-    component = await transaction.get_component(cmd.entity_id, ArchivePasswordComponent)
-    if component:
-        await transaction.remove_component(component)
-        logger.info(f"Removed ArchivePasswordComponent from entity {cmd.entity_id}")
-
-
-@system(on_command=SetArchivePasswordCommand)
-async def set_archive_password_handler(
-    cmd: SetArchivePasswordCommand,
-    transaction: WorldTransaction,
-) -> None:
-    """
-    Handles setting the password for an archive.
-    """
-    password_comp = await transaction.get_component(cmd.entity_id, ArchivePasswordComponent)
-    if password_comp:
-        password_comp.password = cmd.password
-    else:
-        password_comp = ArchivePasswordComponent(password=cmd.password)
-        await transaction.add_component_to_entity(cmd.entity_id, password_comp)
 
 
 @system(on_command=GetAssetStreamCommand)
@@ -465,7 +88,7 @@ async def get_archive_asset_stream_handler(
         # Get the provider for the parent archive stream
         archive_stream_cmd = GetAssetStreamCommand(entity_id=target_entity_id)
         try:
-            archive_stream_provider = await world.dispatch_command(archive_stream_cmd).get_first_non_none_value()
+            archive_stream_provider = await world.dispatch_command(archive_stream_cmd).get_one_value()
         except ValueError:
             archive_stream_provider = None
